Moves slug to a encode/decode rather than relying on entropy for uniqueness.
Can't rely on the fact that a randomly generated string will actually be random. The only way to guarantee uniqueness is to make the slug a mapping function to the ID of the record, because that's guaranteed to be unique.

Admin slug can be the ID slug plus a random string - the ID slug will ensure it's uniqueness, and the random salt will make sure it's not reverse engineerable.